requirements.txt
progress
python-whois